backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-filter==24.3
psycopg2-binary==2.9.10
celery==5.4.0
redis==5.2.1
pandas==2.2.3
openpyxl==3.1.5
numpy==2.1.3
Pillow==11.0.0
gunicorn==23.0.0
python-dotenv==1.0.1
